+++ libraries/exoplayer/src/test/java/androidx/media3/exoplayer/e2etest/WebvttPlaybackTest.java
@@ -22,7 +22,6 @@ import static com.google.common.truth.Truth.assertThat;
 import android.content.Context;
 import android.graphics.SurfaceTexture;
 import android.net.Uri;
-import android.os.Looper;
 import android.view.Surface;
 import androidx.media3.common.C;
 import androidx.media3.common.MediaItem;
@@ -35,11 +34,9 @@ import androidx.media3.exoplayer.DefaultLoadControl;
 import androidx.media3.exoplayer.DefaultRenderersFactory;
 import androidx.media3.exoplayer.ExoPlaybackException;
 import androidx.media3.exoplayer.ExoPlayer;
-import androidx.media3.exoplayer.Renderer;
 import androidx.media3.exoplayer.RenderersFactory;
 import androidx.media3.exoplayer.source.DefaultMediaSourceFactory;
 import androidx.media3.exoplayer.source.MediaSource;
-import androidx.media3.exoplayer.text.TextOutput;
 import androidx.media3.exoplayer.text.TextRenderer;
 import androidx.media3.test.utils.CapturingRenderersFactory;
 import androidx.media3.test.utils.DumpFileAsserts;
@@ -50,8 +47,6 @@ import androidx.media3.test.utils.robolectric.ShadowMediaCodecConfig;
 import androidx.media3.test.utils.robolectric.TestPlayerRunHelper;
 import androidx.test.core.app.ApplicationProvider;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.Iterables;
-import java.util.ArrayList;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicBoolean;
 import org.junit.Rule;
@@ -78,13 +73,9 @@ public class WebvttPlaybackTest {
     Context applicationContext = ApplicationProvider.getApplicationContext();
     CapturingRenderersFactory capturingRenderersFactory =
         new CapturingRenderersFactory(applicationContext);
-    MediaSource.Factory mediaSourceFactory =
-        new DefaultMediaSourceFactory(applicationContext)
-            .experimentalParseSubtitlesDuringExtraction(true);
     ExoPlayer player =
         new ExoPlayer.Builder(applicationContext, capturingRenderersFactory)
             .setClock(new FakeClock(/* isAutoAdvancing= */ true))
-            .setMediaSourceFactory(mediaSourceFactory)
             .build();
     Surface surface = new Surface(new SurfaceTexture(/* texName= */ 1));
     player.setVideoSurface(surface);
@@ -120,13 +111,9 @@ public class WebvttPlaybackTest {
     Context applicationContext = ApplicationProvider.getApplicationContext();
     CapturingRenderersFactory capturingRenderersFactory =
         new CapturingRenderersFactory(applicationContext);
-    MediaSource.Factory mediaSourceFactory =
-        new DefaultMediaSourceFactory(applicationContext)
-            .experimentalParseSubtitlesDuringExtraction(true);
     ExoPlayer player =
         new ExoPlayer.Builder(applicationContext, capturingRenderersFactory)
             .setClock(new FakeClock(/* isAutoAdvancing= */ true))
-            .setMediaSourceFactory(mediaSourceFactory)
             .setLoadControl(
                 new DefaultLoadControl.Builder()
                     .setBackBuffer(
@@ -167,11 +154,21 @@ public class WebvttPlaybackTest {
         applicationContext, playbackOutput, "playbackdumps/webvtt/" + inputFile + ".seek.dump");
   }
 
+  // Using deprecated TextRenderer.experimentalSetLegacyDecodingEnabled() and
+  // MediaSource.Factory.experimentalParseSubtitlesDuringExtraction() methods to ensure legacy
+  // subtitle handling keeps working.
+  @SuppressWarnings("deprecation")
   @Test
   public void test_legacyParseInRenderer() throws Exception {
     Context applicationContext = ApplicationProvider.getApplicationContext();
     CapturingRenderersFactory capturingRenderersFactory =
-        new CapturingRenderersFactory(applicationContext);
+        new CapturingRenderersFactory(applicationContext)
+            .setTextRendererFactory(
+                (textOutput, outputLooper) -> {
+                  TextRenderer renderer = new TextRenderer(textOutput, outputLooper);
+                  renderer.experimentalSetLegacyDecodingEnabled(true);
+                  return renderer;
+                });
     MediaSource.Factory mediaSourceFactory =
         new DefaultMediaSourceFactory(applicationContext)
             .experimentalParseSubtitlesDuringExtraction(false);
@@ -215,11 +212,21 @@ public class WebvttPlaybackTest {
         applicationContext, playbackOutput, "playbackdumps/webvtt/" + inputFile + ".dump");
   }
 
+  // Using deprecated TextRenderer.experimentalSetLegacyDecodingEnabled() and
+  // MediaSource.Factory.experimentalParseSubtitlesDuringExtraction() methods to ensure legacy
+  // subtitle handling keeps working.
+  @SuppressWarnings("deprecation")
   @Test
   public void test_legacyParseInRendererWithSeek() throws Exception {
     Context applicationContext = ApplicationProvider.getApplicationContext();
     CapturingRenderersFactory capturingRenderersFactory =
-        new CapturingRenderersFactory(applicationContext);
+        new CapturingRenderersFactory(applicationContext)
+            .setTextRendererFactory(
+                (textOutput, outputLooper) -> {
+                  TextRenderer renderer = new TextRenderer(textOutput, outputLooper);
+                  renderer.experimentalSetLegacyDecodingEnabled(true);
+                  return renderer;
+                });
     MediaSource.Factory mediaSourceFactory =
         new DefaultMediaSourceFactory(applicationContext)
             .experimentalParseSubtitlesDuringExtraction(false);
@@ -278,22 +285,12 @@ public class WebvttPlaybackTest {
         applicationContext, playbackOutput, "playbackdumps/webvtt/" + inputFile + ".seek.dump");
   }
 
+  // Deliberately configuring legacy subtitle handling to check unconfigured TextRenderer fails.
+  @SuppressWarnings("deprecation")
   @Test
-  public void textRendererDoesntSupportLegacyDecoding_playbackFails() throws Exception {
+  public void textRenderer_doesntSupportLegacyDecodingByDefault_playbackFails() throws Exception {
     Context applicationContext = ApplicationProvider.getApplicationContext();
-    RenderersFactory renderersFactory =
-        new DefaultRenderersFactory(applicationContext) {
-          @Override
-          protected void buildTextRenderers(
-              Context context,
-              TextOutput output,
-              Looper outputLooper,
-              @ExtensionRendererMode int extensionRendererMode,
-              ArrayList<Renderer> out) {
-            super.buildTextRenderers(context, output, outputLooper, extensionRendererMode, out);
-            ((TextRenderer) Iterables.getLast(out)).experimentalSetLegacyDecodingEnabled(false);
-          }
-        };
+    RenderersFactory renderersFactory = new DefaultRenderersFactory(applicationContext);
     MediaSource.Factory mediaSourceFactory =
         new DefaultMediaSourceFactory(applicationContext)
             .experimentalParseSubtitlesDuringExtraction(false);