+++ libraries/exoplayer_dash/src/test/java/androidx/media3/exoplayer/dash/e2etest/DashPlaybackTest.java
@@ -58,9 +58,6 @@ public final class DashPlaybackTest {
     ExoPlayer player =
         new ExoPlayer.Builder(applicationContext, capturingRenderersFactory)
             .setClock(new FakeClock(/* isAutoAdvancing= */ true))
-            .setMediaSourceFactory(
-                new DashMediaSource.Factory(new DefaultDataSource.Factory(applicationContext))
-                    .experimentalParseSubtitlesDuringExtraction(true))
             .build();
     player.setVideoSurface(new Surface(new SurfaceTexture(/* texName= */ 1)));
     PlaybackOutput playbackOutput = PlaybackOutput.register(player, capturingRenderersFactory);
@@ -87,9 +84,6 @@ public final class DashPlaybackTest {
     ExoPlayer player =
         new ExoPlayer.Builder(applicationContext, capturingRenderersFactory)
             .setClock(new FakeClock(/* isAutoAdvancing= */ true))
-            .setMediaSourceFactory(
-                new DashMediaSource.Factory(new DefaultDataSource.Factory(applicationContext))
-                    .experimentalParseSubtitlesDuringExtraction(true))
             .build();
     player.setVideoSurface(new Surface(new SurfaceTexture(/* texName= */ 1)));
     PlaybackOutput playbackOutput = PlaybackOutput.register(player, capturingRenderersFactory);
@@ -116,9 +110,6 @@ public final class DashPlaybackTest {
         new CapturingRenderersFactory(applicationContext);
     ExoPlayer player =
         new ExoPlayer.Builder(applicationContext, capturingRenderersFactory)
-            .setMediaSourceFactory(
-                new DashMediaSource.Factory(new DefaultDataSource.Factory(applicationContext))
-                    .experimentalParseSubtitlesDuringExtraction(true))
             .setClock(new FakeClock(/* isAutoAdvancing= */ true))
             .build();
     player.setVideoSurface(new Surface(new SurfaceTexture(/* texName= */ 1)));
@@ -145,9 +136,6 @@ public final class DashPlaybackTest {
         new CapturingRenderersFactory(applicationContext);
     ExoPlayer player =
         new ExoPlayer.Builder(applicationContext, capturingRenderersFactory)
-            .setMediaSourceFactory(
-                new DashMediaSource.Factory(new DefaultDataSource.Factory(applicationContext))
-                    .experimentalParseSubtitlesDuringExtraction(true))
             .setClock(new FakeClock(/* isAutoAdvancing= */ true))
             .build();
     player.setVideoSurface(new Surface(new SurfaceTexture(/* texName= */ 1)));
@@ -171,6 +159,9 @@ public final class DashPlaybackTest {
    * This test and {@link #cea608_parseDuringExtraction()} use the same output dump file, to
    * demonstrate the flag has no effect on the resulting subtitles.
    */
+  // Using deprecated MediaSource.Factory.experimentalParseSubtitlesDuringExtraction() method to
+  // ensure legacy subtitle handling keeps working.
+  @SuppressWarnings("deprecation")
   @Test
   public void cea608_parseDuringRendering() throws Exception {
     Context applicationContext = ApplicationProvider.getApplicationContext();
@@ -204,6 +195,9 @@ public final class DashPlaybackTest {
    * This test and {@link #cea608_parseDuringRendering()} use the same output dump file, to
    * demonstrate the flag has no effect on the resulting subtitles.
    */
+  // Explicitly enable parsing during extraction (even though a) it's the default and b) currently
+  // all CEA-608 parsing happens during rendering) to make this test clearer & more future-proof.
+  @SuppressWarnings("deprecation")
   @Test
   public void cea608_parseDuringExtraction() throws Exception {
     Context applicationContext = ApplicationProvider.getApplicationContext();